+++ src/pages/add/add.styles.tsx
@@ -3,6 +3,5 @@ import styled from "styled-components";
 export const Title = styled.h2`
   font-size: 18pt;
   margin-bottom: 20px;
-  margin-left: 30px;
   color: #474b4e;
 `;