+++ src/pages/search/search.styles.tsx
@@ -3,7 +3,6 @@ import styled from "styled-components";
 export const Title = styled.h2`
   font-size: 18pt;
   margin-bottom: 20px;
-  margin-left: 30px;
   color: #474b4e;
 `;